Chore: working on location settings

+++ lib/presentation/pages/account/account_page.dart
@@ -15,6 +15,7 @@ import 'package:test_futter_project/presentation/pages/account/widgets/account_i
 import 'package:test_futter_project/presentation/pages/account/widgets/custom_divider.dart';
 import 'package:test_futter_project/presentation/pages/account/widgets/user_avatar.dart';
 import 'package:test_futter_project/presentation/pages/authentication/login_page.dart';
+import 'package:test_futter_project/utils/dialog_helper.dart';
 import 'package:test_futter_project/utils/l10n.dart';
 
 import '../../../common/app_text_styles.dart';
@@ -122,12 +123,22 @@ class AccountPage extends StatelessWidget {
                 Padding(
                   padding: const EdgeInsets.all(AppDimensions.normalS),
                   child: AccountItemSeparated(
-                    onTap: () async {
-                      await context.read<AuthenticationCubit>().deleteAccount(state.email);
+                    title: AppLocalisations.accountItemDeleteAccount,
+                    onTap: () {
+                      DialogHelper.showConfirmationDialog(
+                        context,
+                        title: AppLocalisations.accountItemDeleteAccount,
+                        description: AppLocalisations.deleteAccountDialogDescription,
+                        cancelButtonTitle: AppLocalisations.deleteAccountDialogCancelLabel,
+                        confirmButtonTitle: AppLocalisations.deleteAccountDialogConfirmLabel,
+                        onConfirm: () async {
+                          await context.read<AuthenticationCubit>().deleteAccount(state.email);
 
-                      if (!context.mounted) return;
+                          if (!context.mounted) return;
 
-                      context.read<UserDataCubit>().logOutUser();
+                          context.read<UserDataCubit>().logOutUser();
+                        },
+                      );
                     },
                   ),
                 ),
